Part 23 Complete
Fixed all remaining Datapack errors and the game is now playable. Ores are now generating, and a few items now show up in the Creative Tab and JEI. This build is VERY stable. I am now done porting to Minecraft 1.14 and will work on the changes introduced in the 1.2.1 version.